API Reference

API Reference

Oracle.AbstractOracleValue — Type.

Holds a 1-indexed vector of OraData.

source

Oracle.CommonCreateParams — Type.

Safe version of OraCommonCreateParams

source

Oracle.ConnCreateParams — Type.

Safe version of OraConnCreateParams

source

Oracle.Connection — Type.

Connection handles are used to represent connections to the database. These can be standalone connections created by calling the function dpiConncreate() or acquired from a session pool by calling the function dpiPoolacquireConnection(). They can be closed by calling the function dpiConnclose() or releasing the last reference to the connection by calling the function dpiConnrelease(). Connection handles are used to create all handles other than session pools and context handles.

source

Oracle.EncodingInfo — Type.

Mirrors ODPI-C's OraEncodingInfo struct, but using Julia types.

source

Oracle.ExternOracleValue — Type.

Wraps a OraData handle managed by extern ODPI-C. 1-indexed.

source

Oracle.JuliaOracleValue — Type.

Wraps a OraData handle managed by Julia. 1-indexed.

source

Oracle.OraCommonCreateParams — Method.

This function may be unsafe: safe_params must outlive the OraCommonCreateParams generated by this function.

source

Oracle.OraConnCreateParams — Method.

This function may be unsafe: safe_params must outlive the OraConnCreateParams generated by this function.

source

Oracle.StmtInfo — Type.

High-level version for OraStmtInfo using Bool Julia type.

source

Oracle.Variable — Method.

Variable is a 1-indexed array of OraData.

source

Oracle.define — Method.

define(stmt::QueryStmt, column_position::Integer, variable::Variable)

Defines the variable that will be used to fetch rows from the statement. stmt must be an executed statement.

A Variable v bound to a statement stmt must satisfy:

v.buffer_capacity >= fetch_array_size(stmt)

source

Oracle.execute — Method.

execute(stmt::Stmt; exec_mode::dpiExecMode=ORA_MODE_EXEC_DEFAULT) :: UInt32

Returns the number of columns which are being queried. If the statement does not refer to a query, the value is set to 0.

source

Oracle.fetch — Method.

fetch(stmt::Stmt) :: FetchResult

Fetches a single row from the statement.

source

Oracle.fetch_array_size! — Method.

fetch_array_size!(stmt::Stmt, new_size::Integer)

Sets the array size used for performing fetches. All variables defined for fetching must have this many (or more) elements allocated for them. The higher this value is the less network round trips are required to fetch rows from the database but more memory is also required.

A value of zero will reset the array size to the default value of DPIDEFAULTFETCHARRAYSIZE.

source

Oracle.row_count — Method.

Number of affected rows in a DML statement.

source