API Reference

API Reference

ODPI-C Naming Conventions

All enums, constants and structs in ODPI-C library use the prefix DPI or dpi.

In Oracle.jl, the Julia implementation of these elements use the prefix ORA or Ora.

Examples:

The ODPI-C constant DPI_MODE_AUTH_SYSDBA becomes ORA_MODE_AUTH_SYSDBA in Julia.

The ODPI-C enum dpiAuthMode becomes OraAuthMode in Julia.

The ODPI-C struct dpiTimestamp becomes OraTimestamp in Julia.

All julia structs with prefix Ora are raw wrappers around ODPC-C structs and may contain unsafe attributes.

Safe equivalent Julia structs drop the Ora prefix.

ODPI-C function wrappers have their name preserved, as in dpiContext_create.

Connection

Oracle.Connection — Type.

Connection(user::AbstractString, password::AbstractString, connect_string::AbstractString;
        encoding::AbstractString=DEFAULT_CONNECTION_ENCODING,
        nencoding::AbstractString=DEFAULT_CONNECTION_NENCODING,
        create_mode::Union{Nothing, OraCreateMode}=nothing,
        edition::Union{Nothing, String}=nothing,
        driver_name::Union{Nothing, String}=nothing,
        auth_mode::OraAuthMode=ORA_MODE_AUTH_DEFAULT,
        pool::Union{Nothing, Pool}=nothing
    )

Creates a connection to the Oracle Database.

Connections should always be closed after use by calling Oracle.close.

Example

import Oracle

username = "my_username"
password = "my_password"
connect_string = "//IP_ADDRESS/XE" # a valid Oracle connect string

conn = Oracle.Connection(username, password, connect_string)

# connections should always be closed after use.
Oracle.close(conn)

source

Oracle.ping — Function.

ping(conn::Connection)

Pings the database server to check if the connection is still alive. Throws error if can't ping the server.

source

Oracle.commit — Function.

commit(conn::Connection)

Commits the current active transaction.

source

Oracle.rollback — Function.

rollback(conn::Connection)

Rolls back the current active transaction.

source